refactor: :recycle: Refactor after PR review, mainly spacing issues and portal fixes.
