+++ packages/shared/package.json
@@ -20,7 +20,7 @@
     "idb-keyval": "^5.0.6",
     "next": "^10.2.3",
     "react": "^17.0.5",
-    "react-contexify": "^5.0.0",
+    "@rebelchris/react-contexify": "^5.0.0",
     "react-dom": "^17.0.5",
     "react-intersection-observer": "^8.31.0",
     "react-modal": "^3.14.3",
@@ -86,7 +86,7 @@
     "postcss-simple-vars": "^6.0.3",
     "prettier": "^2.4.1",
     "react": "^17.0.2",
-    "react-contexify": "^5.0.0",
+    "@rebelchris/react-contexify": "^5.0.0",
     "react-dom": "^17.0.2",
     "react-intersection-observer": "^8.32.1",
     "react-modal": "^3.14.3",
