+++ packages/shared/src/components/modals/ReportPostModal.tsx
@@ -3,9 +3,9 @@ import { Radio } from '../fields/Radio';
 import { Post } from '../../graphql/posts';
 import {
   ConfirmationModal,
-  ConfirmationHeading,
   ConfirmationDescription,
   ConfirmationButtonsCenter,
+  BigConfirmationHeading,
 } from './ConfirmationModal';
 import { ModalCloseButton } from './ModalCloseButton';
 import { Checkbox } from '../fields/Checkbox';
@@ -44,7 +44,7 @@ export default function RepostPostModal({
   return (
     <ConfirmationModal {...props}>
       <ModalCloseButton onClick={props.onRequestClose} />
-      <ConfirmationHeading>Report article</ConfirmationHeading>
+      <BigConfirmationHeading>Report article</BigConfirmationHeading>
       <ConfirmationDescription>
         &quot;{post?.title}&quot;
       </ConfirmationDescription>
@@ -56,26 +56,26 @@ export default function RepostPostModal({
           value={reason}
           onChange={setReason}
         />
-        <div className="-mx-10 h-px bg-theme-divider-secondary" />
+        <div className="-mx-8 h-px bg-theme-divider-secondary" />
         <p className="mt-6 mb-1 font-bold typo-caption1">
           Anything else you&apos;d like to add?
         </p>
         <textarea
           onChange={(event) => setComment(event.target.value)}
-          className="self-stretch py-2 px-4 mt-1 mb-6 w-full h-20 bg-theme-float rounded-10 resize-none typo-body"
+          className="self-stretch p-2 mb-4 w-full h-20 bg-theme-float rounded-10 resize-none typo-body"
         />
         <Checkbox
           ref={inputRef}
           name="blockSource"
-          className="self-center mb-6 font-normal"
+          className="self-center mx-2 mb-6 font-normal"
         >
-          Don&apos;t show articles from {post?.source?.name}
+          {`Don't show articles from ${post?.source?.name}`}
         </Checkbox>
       </div>
       <ConfirmationButtonsCenter>
         <Button
           className="btn-primary"
-          disabled={!reason}
+          disabled={!reason || (reason === 'OTHER' && !comment)}
           onClick={onReportPost}
         >
           Submit report
